Refactor CustomRoute to use hooks
